Changes from the review
The "Purchase Album" link shows a color contrast warning
Hover styles of links in headers are lacking or insufficient
One set of matching for and id attributes has an illegal space
One fieldset doesn't have a legend tag
Double check your aria role attributes to make sure they are valid
Also, I noticed that on line 29 of your tour.html file you are using a
heading tag, but heading tags are actually defined like this, h1 or h2
and so on down to h6. That's worth changing.